feat: move TitleBar to ui module
