+++ client/app/lib/kivy_manager.py
@@ -3,19 +3,17 @@ import json
 from typing import Any, TypedDict
 from uuid import UUID
 
-import kivymd.uix.button
 import websockets
+from app import ui
 from kivy import Logger
 from kivy.core.window import Window
 from kivy.lang.builder import Builder
 from kivy.modules import inspector
 from kivy.properties import BooleanProperty, StringProperty
-from kivy.utils import get_color_from_hex
 from kivymd.app import MDApp
 from kivymd.uix.boxlayout import MDBoxLayout
-from kivymd.uix.floatlayout import MDFloatLayout
 
-from ..utils import Colors, app_dir
+from ..utils import app_dir
 
 Window.borderless = True
 Window.custom_titlebar = True
@@ -32,38 +30,6 @@ class KivyIds(TypedDict):
     chats_screen_manager: str
     username: str
     password: str
-
-
-class TitleBar(MDFloatLayout):
-    """Custom TitleBar for the app"""
-
-    md_bg_color = get_color_from_hex(Colors.accent_bg.value)
-    button_bg = get_color_from_hex(Colors.primary_bg.value)
-    button_size = "15sp"
-
-    def __init__(self, **kwargs):
-        super(TitleBar, self).__init__(**kwargs)
-        self.app: MDApp = MDApp.get_running_app()
-
-    def handle_buttons(self, instance: kivymd.uix.button.BaseButton):
-        """Callback Function for all buttons in titlebar
-
-        :param instance a button object that is a subclass of kivymd.uix.button.BaseButton
-
-        A button object was chosen instead of say a string, so that later on the caller itself can be edited
-        """
-        match_string = None
-        if hasattr(instance, "icon"):
-            match_string = instance.icon
-
-        if match_string:
-            match match_string:
-                case "close":
-                    self.app.stop()
-                case "window-minimize":
-                    self.app.root_window.minimize()
-                case "window-maximize":
-                    self.app.root_window.maximize()
 
 
 class ClientUI(MDApp):
@@ -96,7 +62,7 @@ class ClientUI(MDApp):
             Builder.load_file(str(file))
         # load root explicitly
         root = Builder.load_file(str(app_dir / "ui/kv_files/client_ui.kv"))
-        root.md_bg_color = get_color_from_hex(Colors.primary_bg.value)
+        root.ids["titlebar"]: ui.TitleBar
         if Window.set_custom_titlebar(root.ids["titlebar"]):
             Logger.info("Window: setting custom titlebar successful")
         else:
