+++ client/app/ui/__init__.py
@@ -1,5 +1,10 @@
+from kivy.utils import get_color_from_hex
 from kivymd.app import MDApp
+from kivymd.uix.button import BaseButton
 from kivymd.uix.card import MDCard
+from kivymd.uix.floatlayout import MDFloatLayout
+
+from ..utils import Colors
 
 
 class ChatItem(MDCard):
@@ -28,3 +33,35 @@ class ChatItem(MDCard):
             MDApp.get_running_app().root.ids[
                 "screen_manager"
             ].current = self.custom_id  # switch screen to the chat
+
+
+class TitleBar(MDFloatLayout):
+    """Custom TitleBar for the app"""
+
+    md_bg_color = get_color_from_hex(Colors.accent_bg.value)
+    button_bg = get_color_from_hex(Colors.primary_bg.value)
+    button_size = "15sp"
+
+    def __init__(self, **kwargs):
+        super(TitleBar, self).__init__(**kwargs)
+        self.app: MDApp = MDApp.get_running_app()
+
+    def handle_buttons(self, instance: BaseButton):
+        """Callback Function for all buttons in titlebar
+
+        :param instance a button object that is a subclass of kivymd.uix.button.BaseButton
+
+        A button object was chosen instead of say a string, so that later on the caller itself can be edited
+        """
+        match_string = None
+        if hasattr(instance, "icon"):
+            match_string = instance.icon
+
+        if match_string:
+            match match_string:
+                case "close":
+                    self.app.stop()
+                case "window-minimize":
+                    self.app.root_window.minimize()
+                case "window-maximize":
+                    self.app.root_window.maximize()
